requirements.txt
flask==1.1.2
pytest==6.2.3
invoke==1.5.0
<<<<<<< HEAD
pycodestyle = 2.7.0
pylint = 2.8.3
pytest-cov = 2.12.1
=======
pycodestyle==2.7.0
pylint==2.8.3
>>>>>>> 0da701fb5549b90088f5c061ba5b3cd2aa35056f

pyyaml==5.4.1
xmltodict==0.12.0